Enable interactive component (#38)

+++ diffusion2/app_interactive.py
@@ -3,13 +3,13 @@
 # !curl https://raw.githubusercontent.com/Lightning-AI/stablediffusion/main/configs/stable-diffusion/v2-inference-v.yaml -o v2-inference-v.yaml
 import lightning as L
 import os
-import asyncio, torch, base64, functools
+import inspect
+import asyncio, torch, base64, functools, time
 from io import BytesIO
 from typing import Any, Callable, Optional
 from ldm.lightning import LightningStableDiffusion, PromptDataset
 from nicegui import ui
 from pydantic import BaseModel
-
 
 os.environ["PYTORCH_ENABLE_MPS_FALLBACK"] = "1"
 
@@ -25,14 +25,25 @@ async def io_bound(callback: Callable, *args: Any, **kwargs: Any):
 
 
 def webpage(
-    predict_fn: Callable, host: Optional[str] = None, port: Optional[int] = None
-):
+    predict_fn: Callable, host: str, port: int, reference_inference_time: Optional[float]
+):    
+
+    async def progress_tracker():
+        if progress.value >= 1.0 or progress.value == 0:
+            return
+        progress.value = round((progress.value * reference_inference_time + 0.1) / reference_inference_time, 3)
+
     async def generate_image():
+        nonlocal reference_inference_time
+        t0 = time.time()
+        progress.value = 0.0001
         image.source = (
             "https://dummyimage.com/600x400/ccc/000000.png&text=building+image..."
         )
         prediction = await io_bound(predict_fn, request=Text(text=prompt.value))
         image.source = prediction["image"]
+        progress.value = 1.0
+        reference_inference_time = (time.time() - t0)
 
     # User Interface
     with ui.row().style("gap:10em"):
@@ -41,31 +52,34 @@ def webpage(
             prompt = ui.input("prompt").style("width: 50em")
             prompt.value = "Woman painting a large red egg in a dali landscape"
             ui.button("Generate", on_click=generate_image).style("width: 15em")
+            progress = ui.linear_progress()
+            ui.timer(interval=0.1, callback=progress_tracker)
             image = ui.image().style("width: 60em")
 
+    # Note: Hack to enable running in spawn context.
+    def stack_patch():
+        class FakeFrame:
+            filename = "random"
+        return [FakeFrame(), None]
+    inspect.stack = stack_patch
     ui.run(host=host, port=port, reload=False)
 
 
 class DiffusionServeInteractive(L.LightningWork):
+
+    _start_method = "spawn"
+
     def setup(self):
-        os.system(
-            "curl https://pl-public-data.s3.amazonaws.com/dream_stable_diffusion/768-v-ema.ckpt -o 768-v-ema.ckpt"
-        )
-        os.system("echo *.ckpt > .lightningignore ")
-
-        running_local = os.getenv("LIGHTNING_CLOUD_APP_ID", None) is None
-
-        if running_local:
-            precision = 32
-            accelerator = "cpu"
-        else:
-            precision = 16 if torch.cuda.is_available() else 32
-            accelerator = "auto"
+        if not os.path.exists("768-v-ema.ckpt"):
+            os.system(
+                "curl https://pl-public-data.s3.amazonaws.com/dream_stable_diffusion/768-v-ema.ckpt -o 768-v-ema.ckpt"
+            )
+            os.system("echo *.ckpt > .lightningignore")
 
         self._trainer = L.Trainer(
-            accelerator=accelerator,
+            accelerator="auto",
             devices=1,
-            precision=precision,
+            precision=16 if torch.cuda.is_available() else 32,
             enable_progress_bar=False,
         )
 
@@ -76,7 +90,7 @@ class DiffusionServeInteractive(L.LightningWork):
             size=768,
         )
 
-        if not running_local and torch.cuda.is_available():
+        if torch.cuda.is_available():
             self._model = self._model.to(torch.float16)
             torch.cuda.empty_cache()
 
@@ -89,10 +103,16 @@ class DiffusionServeInteractive(L.LightningWork):
         image.save(buffer, format="PNG")
         img_str = base64.b64encode(buffer.getvalue()).decode("utf-8")
         return {"image": f"data:image/png;base64,{img_str}"}
+         
+    def warm_up(self):
+        self._model.steps = 1
+        self.predict(Text(text='Warm-up Machine'))
+        self._model.steps = 50
 
     def run(self):
         self.setup()
-        webpage(self.predict, host=self.host, port=self.port)
+        self.warm_up()
+        webpage(self.predict, host=self.host, port=self.port, reference_inference_time=15)
 
 
 component = DiffusionServeInteractive(
